+++ src/main/kotlin/net/pwall/json/schema/codegen/Validation.kt
@@ -2,7 +2,7 @@
  * @(#) Validation.kt
  *
  * json-kotlin-schema-codegen  JSON Schema Code Generation
- * Copyright (c) 2020, 2023 Peter Wall
+ * Copyright (c) 2020, 2023, 2024 Peter Wall
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -48,7 +48,7 @@ class Validation(
         MINIMUM_DECIMAL_ZERO, EXCLUSIVE_MINIMUM_DECIMAL_ZERO,
         RANGE_INT, RANGE_LONG, RANGE_DECIMAL,
         EMAIL, HOSTNAME, IPV4, IPV6, DURATION, JSON_POINTER, RELATIVE_JSON_POINTER,
-        DATE_TIME, DATE, TIME, UUID, URI, URI_REFERENCE,
+        DATE_TIME, DATE, TIME, UUID, URI, URI_REFERENCE, URI_TEMPLATE,
         MAX_ITEMS, MIN_ITEMS, CONST_ITEMS, RANGE_ITEMS,
         MAX_LENGTH, MIN_LENGTH, CONST_LENGTH, RANGE_LENGTH,
         CONST_INT, CONST_LONG, CONST_DECIMAL, CONST_DECIMAL_ZERO, CONST_STRING,