+++ src/test/kotlin/net/pwall/json/schema/codegen/CodeGeneratorUtilTest.kt
@@ -2,7 +2,7 @@
  * @(#) CodeGeneratorUtilTest.kt
  *
  * json-kotlin-schema-codegen  JSON Schema Code Generation
- * Copyright (c) 2023 Peter Wall
+ * Copyright (c) 2023, 2024 Peter Wall
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -30,6 +30,7 @@ import kotlin.test.assertSame
 import kotlin.test.expect
 
 import net.pwall.json.schema.codegen.CodeGenerator.Companion.sanitiseName
+import net.pwall.json.schema.codegen.CodeGenerator.Companion.toColumnId
 
 class CodeGeneratorUtilTest {
 
@@ -45,6 +46,21 @@ class CodeGeneratorUtilTest {
     @Test fun `should sanitise name correctly`() {
         expect("Abc123") { "Abc-123".sanitiseName() }
         expect("Abc123456") { "Abc-123-456".sanitiseName() }
+    }
+
+    @Test fun `should convert number to column id`() {
+        expect("A") { 0.toColumnId() }
+        expect("B") { 1.toColumnId() }
+        expect("Z") { 25.toColumnId() }
+        expect("AA") { 26.toColumnId() }
+        expect("AB") { 27.toColumnId() }
+        expect("AZ") { 51.toColumnId() }
+        expect("BA") { 52.toColumnId() }
+        expect("BB") { 53.toColumnId() }
+        expect("ZA") { (26 * 26).toColumnId() }
+        expect("ZZ") { (26 * 26 + 25).toColumnId() }
+        expect("AAA") { (27 * 26).toColumnId() }
+        expect("AAB") { (27 * 26 + 1).toColumnId() }
     }
 
 }